Loader: Remove obsolete API.
