+++ docs/api/en/loaders/Loader.html
@@ -10,58 +10,6 @@
 	<body>
 		<h1>[name]</h1>
 
-		<p class="desc">Base class for implementing loaders.</p>
-
-
-		<h2>Constructor</h2>
-
-
-		<h3>[name]()</h3>
-		<p>
-		Creates a new [name]. This should be called as base class.
-		</p>
-
-
-		<h2>Properties</h2>
-
-		<h3>[property:Function onLoadStart]</h3>
-		<p>Will be called when load starts.</p>
-		<p>The default is a function with empty body.</p>
-
-		<h3>[property:Function onLoadProgress]</h3>
-		<p>Will be called while load progresses.</p>
-		<p>The default is a function with empty body.</p>
-
-		<h3>[property:Function onLoadComplete]</h3>
-		<p>Will be called when load completes.</p>
-		<p>The default is a function with empty body.</p>
-
-		<h3>[property:string crossOrigin]</h3>
-		<p>
-		The crossOrigin string to implement CORS for loading the url from a different domain that allows CORS.
-		Default is *"anonymous"*.
-		</p>
-
-		<h2>Methods</h2>
-
-		<h3>[method:Material createMaterial]( [param:object m], [param:string texturePath] )</h3>
-		<p>
-		[page:Object m] — The parameters to create the material. <br />
-		[page:String texturePath] — The base path of the textures.
-		</p>
-		<p>
-		Creates the Material based on the parameters m.
-		</p>
-
-		<h3>[method:Array initMaterials]( [param:Array materials], [param:string texturePath] )</h3>
-		<p>
-		[page:Array materials] — an array of parameters to create materials. <br />
-		[page:String texturePath] —  The base path of the textures.
-		</p>
-		<p>
-		Creates an array of [page:Material] based on the array of parameters m. The index of the parameters decide the correct index of the materials.
-		</p>
-
 		<h2>Handlers</h2>
 
 		<p>
